Abstract out common implementation of `BeatSnapGrid`
